requirements.txt
meshtastic
kaspa
sentence-transformers
faiss-cpu
nltk
requests
pycryptodomex
pytest
pytest-asyncio
